+++ spring-security/src/test/java/ru/otus/controller/BookControllerTest.java
@@ -30,6 +30,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 
 @DisplayName("Класс BookControllerTest ")
 @WebMvcTest(BookController.class)
+@WithMockUser
 public class BookControllerTest {
 
     @Autowired
@@ -47,10 +48,8 @@ public class BookControllerTest {
     @MockBean
     private GenreService genreService;
 
-    private final long bookId = 1;
 
     @DisplayName("возвращает все книги")
-    @WithMockUser(username = "user")
     @Test
     void booksPageShouldReturnCorrectView() throws Exception {
         Author author = new Author(1,"Test_author");
@@ -67,7 +66,6 @@ public class BookControllerTest {
     }
 
     @DisplayName("возвращает книгу со всеми комментариями")
-    @WithMockUser(username = "user")
     @Test
     void bookPageShouldReturnCorrectView() throws Exception {
         Author author = new Author(1,"Test_author");
@@ -87,7 +85,6 @@ public class BookControllerTest {
     }
 
     @DisplayName("возвращает страницу для редактирования книги")
-    @WithMockUser(username = "user")
     @Test
     void editPageShouldReturnCorrectView() throws Exception {
         Author author1 = new Author(1,"Test_author1");
@@ -109,8 +106,8 @@ public class BookControllerTest {
                 .andExpect(model().attribute("genres", genreList));
     }
 
+
     @DisplayName("сохраняет книгу")
-    @WithMockUser(username = "user")
     @Test
     void saveBookShouldReturnCorrectView() throws Exception {
         Author author = new Author(1,"Test_author");
@@ -125,7 +122,6 @@ public class BookControllerTest {
 
 
     @DisplayName("возвращает страницу для создания новой книги")
-    @WithMockUser(username = "user")
     @Test
     void createBookPageShouldReturnCorrectView() throws Exception {
         Author author1 = new Author(1,"Test_author1");
@@ -145,7 +141,6 @@ public class BookControllerTest {
     }
 
     @DisplayName("создаёт новую книгу")
-    @WithMockUser(username = "user")
     @Test
     void createBookShouldReturnCorrectView() throws Exception {
         long bookId = 1;
@@ -159,7 +154,6 @@ public class BookControllerTest {
     }
 
     @DisplayName("удаляет книгу")
-    @WithMockUser(username = "user")
     @Test
     void deleteBookShouldReturnCorrectView() throws Exception {
         long bookId = 1;
@@ -170,62 +164,6 @@ public class BookControllerTest {
         mvc.perform(delete("/book/" + bookId).with(csrf()))
                 .andExpect(status().is3xxRedirection());
         assertEquals(bookId, valueCapture.getValue());
-    }
-
-    @DisplayName("неавторизированным пользователям разрешает доступ к странице логина")
-    @Test
-    void givenAccessToLoginPage() throws Exception {
-        mvc.perform(get("/login"))
-                .andExpect(status().isOk());
-    }
-
-    @DisplayName("неавторизированным пользователям запрещает доступ к корневой странице")
-    @Test
-    void forbiddenBooksPage() throws Exception {
-        mvc.perform(get("/"))
-                .andExpect(status().isUnauthorized());
-    }
-
-    @DisplayName("неавторизированным пользователям запрещает доступ к странице книги")
-    @Test
-    void forbiddenBookPage() throws Exception {
-        mvc.perform(get("/book/" + bookId))
-                .andExpect(status().isUnauthorized());
-    }
-
-    @DisplayName("неавторизированным пользователям запрещает доступ к странице для редактирования книги")
-    @Test
-    void forbiddenBookEditPage() throws Exception {
-        mvc.perform(get("/edit?id=" + bookId))
-                .andExpect(status().isUnauthorized());
-    }
-
-    @DisplayName("неавторизированным пользователям запрещает сохранять книгу")
-    @Test
-    void forbiddenToSaveBook() throws Exception {
-        mvc.perform(put( "/edit").with(csrf()))
-                .andExpect(status().isUnauthorized());
-    }
-
-    @DisplayName("неавторизированным пользователям запрещает возвращать страницу для создания новой книги")
-    @Test
-    void forbiddenToReturnCreateBookPage() throws Exception {
-        mvc.perform(get("/book"))
-                .andExpect(status().isUnauthorized());
-    }
-
-    @DisplayName("неавторизированным пользователям запрещает создавать новую книгу")
-    @Test
-    void forbiddenToCreateBook() throws Exception {
-        mvc.perform(post("/book").with(csrf()))
-                .andExpect(status().isUnauthorized());
-    }
-
-    @DisplayName("неавторизированным пользователям запрещает удалять книгу")
-    @Test
-    void forbiddenToDeleteBook() throws Exception {
-        mvc.perform(delete("/book/" + bookId).with(csrf()))
-                .andExpect(status().isUnauthorized());
     }
 }
